requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.8.2
aiofiles==24.1.0
httpx==0.27.2
 